Terrabot

+++ src/main/java/main/Entities/Plants/PlantTypes.java
@@ -1,27 +1,27 @@
 package main.Entities.Plants;
 
 enum PlantTypes {
-    FLOWERING(6.0f, 90, "FloweringPlants"),
-    GYMNOSPERM(0.0f, 60, "GymnospermsPlants"),
-    FERN(0.0f, 30, "Ferns"),
-    MOSS(0.8f, 40, "Mosses"),
-    ALGA(0.5f, 20, "Algae");
+    FLOWERING(6.0, 90, "FloweringPlants"),
+    GYMNOSPERM(0.0, 60, "GymnospermsPlants"),
+    FERN(0.0, 30, "Ferns"),
+    MOSS(0.8, 40, "Mosses"),
+    ALGA(0.5, 20, "Algae");
 
-    private float baseOxygen;
-    private float tangleChance;
+    private double baseOxygen;
+    private double tangleChance;
     private String typeName;
 
-    PlantTypes(float baseOxygen, int tanglePossiblity, String typeName) {
+    PlantTypes(double baseOxygen, int tanglePossiblity, String typeName) {
         this.baseOxygen = baseOxygen;
-        this.tangleChance = (float) tanglePossiblity / 100.0f;
+        this.tangleChance = tanglePossiblity / 100.0;
         this.typeName = typeName;
     }
 
-    public float getBaseOxygen() {
+    public double getBaseOxygen() {
         return baseOxygen;
     }
 
-    public float getTangleChance() {
+    public double getTangleChance() {
         return tangleChance;
     }
